PycharmProjects/yohire2/requirements.txt
python-dotenv==1.1.0
# Загружает пары ключ=значение из файла .env в переменные окружения (os.environ).
# Чтобы не хардкодить пароли и креды в коде, а держать их в одном конфиденциальном файле.

fastapi==0.115.12
# Современный, быстрый (асинхронный) веб-фреймворк на Python для написания API.
# Работает через типы (type hints), автоматом генерирует документацию Swagger, легко расширяется.

sqlalchemy==2.0.41
# Мощный SQL-toolkit и ORM. Позволяет описывать модели таблиц как классы Python
# Cтроить запросы из кода, при этом под капотом генерируя чистый SQL.

uvicorn==0.34.3
# Лёгкий ASGI-сервер, на котором запускается FastAPI. Обрабатывает HTTP/1.1 и WebSocket, умеет делать “auto-reload”

psycopg2-binary==2.9.10
# Бинарный драйвер для PostgreSQL. Дают готовую сборку с C-расширениями,
# чтобы подключаться к Postgres и делать SQL-запросы.